Fix error "undefined method `+' for nil:NilClass" for Linux
A multi-platform cookbook can depends on notepadpp for all the
platforms it supports, since it is not possible to declare windows
only dependencies.

But in this case, notepadpp cookbook raises an error when platform
not Windows (eg Linux) when attributes/default.rb is used. It's
because ENV['ProgramFiles'] is Nil.

This fix adds 'return' early when platform is not Windows.